Rename CRUKSearchSummary to CRUKSearchkitSummary.
