+++ src/styles/common.scss
@@ -88,7 +88,7 @@ $react-dates-color-text-focus: #ffffff;
 @import './../components/loading/loading';
 @import './../components/pagination/CRUKSearchkitPagination';
 @import './../components/result/CRUKSearchkitResult';
-@import './../components/summary/summary';
+@import './../components/summary/CRUKSearchSummary';
 @import './../components/ui/search/CRUKSearchkitSearchUI';
 @import './../components/date/cruk_daterange';
 @import './../components/geosuggest/cruk_geosuggest';
